+++ src/app/[locale]/search/components/SearchFilters.tsx
@@ -51,7 +51,7 @@ export function SearchFilters({
     <div className="flex flex-col gap-8">
       <div className="flex max-w-3xl flex-col gap-3 sm:flex-row">
         <div className="relative flex-1">
-          <Search className="absolute top-1/2 left-4 h-5 w-5 -translate-y-1/2 text-[#d4af37]" />
+          <Search className="absolute top-1/2 left-4 size-5 -translate-y-1/2 text-[#d4af37]" />
           <Input
             value={query}
             onChange={(e) => onQueryChange(e.target.value)}
@@ -61,16 +61,17 @@ export function SearchFilters({
           />
           {query && (
             <button
+              type="button"
               onClick={() => onQueryChange('')}
               className="absolute top-1/2 right-4 -translate-y-1/2 text-[#a48a90] transition-colors hover:text-[#dae2fd]"
             >
-              <X className="h-5 w-5" />
+                  <X className="size-5" />
             </button>
           )}
         </div>
 
         <div className="relative sm:w-48">
-          <MapPin className="absolute top-1/2 left-4 z-10 h-5 w-5 -translate-y-1/2 text-[#d4af37]" />
+          <MapPin className="absolute top-1/2 left-4 z-10 size-5 -translate-y-1/2 text-[#d4af37]" />
           <Select value={selectedCity || '_all'} onValueChange={handleCityChange}>
             <SelectTrigger
               size="xl"
@@ -85,7 +86,7 @@ export function SearchFilters({
               {cities.map((city) => (
                 <SelectItem
                   key={city.slug}
-                  value={city.name}
+                  value={city.slug}
                   className="text-[#dae2fd] focus:bg-[#8b1a4a]/30 focus:text-[#dae2fd]"
                 >
                   {city.name}